# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Fugue API
  version: 1.0.0
  description: |-
    Authoritative API contract for Fugue.
    Route registration, served OpenAPI artifacts, and generated checks all derive from this file.
servers:
  - url: /
...
paths:
...
  /v1/cluster/events:
    get:
      operationId: listClusterEvents
      summary: List Cluster Events
      tags: [Cluster]
      x-fugue-handler: handleListClusterEvents
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/NamespaceQueryParam"
        - $ref: "#/components/parameters/InvolvedObjectKindQueryParam"
        - $ref: "#/components/parameters/InvolvedObjectNameQueryParam"
        - $ref: "#/components/parameters/EventTypeQueryParam"
        - $ref: "#/components/parameters/EventReasonQueryParam"
        - $ref: "#/components/parameters/LimitQueryParam"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterEventListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
  /v1/runtimes/{id}:
    get:
      operationId: getRuntime
      summary: Get Runtime
      tags: [Runtimes]
      x-fugue-handler: handleGetRuntime
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      operationId: deleteRuntime
      summary: Delete Runtime
      tags: [Runtimes]
      x-fugue-handler: handleDeleteRuntime
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteRuntimeResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    RuntimeBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
    EdgeToken:
      type: apiKey
      in: query
      name: token
    DownloadToken:
      type: apiKey
      in: query
      name: download_token
  parameters:
...
    NamespaceQueryParam:
      in: query
      name: namespace
      required: false
      schema:
        type: string
...
    InvolvedObjectKindQueryParam:
      in: query
      name: kind
      required: false
      schema:
        type: string
    InvolvedObjectNameQueryParam:
      in: query
      name: name
      required: false
      schema:
        type: string
    EventTypeQueryParam:
      in: query
      name: type
      required: false
      schema:
        type: string
    EventReasonQueryParam:
      in: query
      name: reason
      required: false
      schema:
        type: string
    LimitQueryParam:
      in: query
      name: limit
      required: false
      schema:
        type: integer
        format: int32
...
  schemas:
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
    StringMap:
      type: object
      additionalProperties:
        type: string
...
    BillingResourceSpec:
      type: object
      additionalProperties: false
      properties:
        cpu_millicores:
          type: integer
          format: int64
        memory_mebibytes:
          type: integer
          format: int64
        storage_gibibytes:
          type: integer
          format: int64
...
    BillingPriceBook:
      type: object
      additionalProperties: false
      required:
        [currency, hours_per_month, cpu_microcents_per_millicore_hour, memory_microcents_per_mib_hour, storage_microcents_per_gib_hour]
      properties:
        currency:
          type: string
        hours_per_month:
          type: integer
          format: int64
        cpu_microcents_per_millicore_hour:
          type: integer
          format: int64
        memory_microcents_per_mib_hour:
          type: integer
          format: int64
        storage_microcents_per_gib_hour:
          type: integer
          format: int64
...
    Runtime:
      type: object
      additionalProperties: false
      required: [id, name, type, status, created_at, updated_at]
      properties:
        id:
          type: string
        tenant_id:
          type: string
        name:
          type: string
        machine_name:
          type: string
        type:
          type: string
        access_mode:
          type: string
        public_offer:
          $ref: "#/components/schemas/RuntimePublicOffer"
        pool_mode:
          type: string
        connection_mode:
          type: string
        status:
          type: string
        endpoint:
          type: string
        labels:
          $ref: "#/components/schemas/StringMap"
        node_key_id:
          type: string
        cluster_node_name:
          type: string
        fingerprint_prefix:
          type: string
        fingerprint_hash:
          type: string
        agent_key_prefix:
          type: string
        agent_key_hash:
          type: string
        last_seen_at:
          type: string
          format: date-time
        last_heartbeat_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    RuntimePublicOffer:
      type: object
      additionalProperties: false
      required: [reference_bundle, price_book, updated_at]
      properties:
        reference_bundle:
          $ref: "#/components/schemas/BillingResourceSpec"
        reference_monthly_price_microcents:
          type: integer
          format: int64
        free:
          type: boolean
        free_cpu:
          type: boolean
        free_memory:
          type: boolean
        free_storage:
          type: boolean
        price_book:
          $ref: "#/components/schemas/BillingPriceBook"
        updated_at:
          type: string
          format: date-time
...
    ClusterEvent:
      type: object
      additionalProperties: false
      required: [namespace, name, type, reason, message, object_kind, object_name]
      properties:
        namespace:
          type: string
        name:
          type: string
        type:
          type: string
        reason:
          type: string
        message:
          type: string
        object_kind:
          type: string
        object_name:
          type: string
        object_namespace:
          type: string
        count:
          type: integer
          format: int32
        first_timestamp:
          type: string
          format: date-time
        last_timestamp:
          type: string
          format: date-time
        event_time:
          type: string
          format: date-time
    ClusterEventListResponse:
      type: object
      additionalProperties: false
      required: [events]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/ClusterEvent"
...
    DeleteRuntimeResponse:
      type: object
      additionalProperties: false
      required: [deleted, runtime]
      properties:
        deleted:
          type: boolean
        runtime:
          $ref: "#/components/schemas/Runtime"
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"